Implemented Lexer and added Token representation that uses enum to identify the type of token and potentially its "value": number, identifier name, etc.
Not sure how we plan to pass tokens to the parser yet, so lexer instance
currently takes a filename and builds a Token list (ordered) internally.
Keyword tokens are matched directly from string to the enum with
toUpperCase() and valueOf(). Punctuation and operators are matched from
a static map in the Lexer class.